ai-api/requirements.txt
fastapi
uvicorn
numpy
opencv-python
keras
tensorflow
mysql-connector-python
python-dotenv
pymongo
gradio==2.0.8